Create an es6 module for the app module
This way we get mor DRY.
